requirements.txt
openai
pylint
pytest
faiss-cpu
sentence-transformers
numpy
python-dotenv
